Fix warning.

+++ tengine/physics/tphysics/tphysics.cpp
@@ -225,7 +225,7 @@ void CTPhysics::Simulate()
 			CTraceResult tr;
 			DetectCollisions(pEntity, mOldPhysics, mNewPhysics, tr);
 
-			for (int i = 0; i < tr.m_aHits.size(); i++)
+			for (size_t i = 0; i < tr.m_aHits.size(); i++)
 			{
 				auto& oOther = tr.m_aHits[i];
